[feat] Open API 데이터베이스 저장 로직 재정립 및 엔드포인트 MovieSearchList By 검색어 개발.

+++ kickmovie/src/main/java/kh/gangnam/kickmovie/entity/MovieDetail.java
@@ -24,14 +24,14 @@ public class MovieDetail {
     @Column(columnDefinition = "TEXT")
     private String overview;    private Double popularity;
     private String releaseDate;
-    private int revenue;
-    private int runtime;
+    private long revenue;
+    private long runtime;
     private String status;
     private String tagline;
     private String title;
     private boolean video;
     private Double voteAverage;
-    private int voteCount;
+    private long voteCount;
 
     @OneToOne
     @JoinColumn(name = "id")
